Add files via upload
A Jupyter Notebook used to try out various different regression models using Scikit-learn. 

The target variable is the price of pumpkins and the independent variables include; month, day of year, city, and variety of pumpkin. The project starts by only using day of year as a regressor and the model fits poorly. Then, a polynomial model was fitted that had an even worse fit. Finally, a model that included dummy variables of the variety variable was used and obtained a determination statistic close to 1 indicating a good fit. However, this model is over-parameterised which causes bias in the model.